ai_insight_generator/pyproject.toml
[project]
name = "ai_insight_generator"
version = "0.1.0"
description = "A platform to transform the 'Reskilling for the AI Economy' book into a live, interactive experience."
authors = [{ name = "Jules", email = "jules@example.com" }]
requires-python = ">=3.9"
dependencies = [
    "pinecone-client>=3.0.0,<4.0.0",
    "sentence-transformers>=2.2.0,<3.0.0",
    "torch>=2.0.0,<3.0.0",
    "tqdm>=4.62.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "numpy",
]
